requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
Werkzeug==3.0.1
openpyxl==3.1.2 

humanize==4.9.0
qrcode==7.4.2
Pillow==10.3.0
APScheduler==3.10.4
ReportLab==4.1.0
PyJWT==2.8.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
pytz==2024.1
